requirements.txt
azure-cosmos==4.6.0
black==24.3.0
python-dotenv==1.0.1